restore old cloud functions
