--- backend/internal/asyncfunc/doc.go
@@ -1,5 +1,0 @@
-/*
-Package asyncfunc provides asynchronous service function logic. Anything that would handle an asynchronous queue message.
-There's almost certainly a better name for this package than this.
-*/
-package asyncfunc
